requirements.txt
numpy==1.14.0
pandas==0.22.0
Pillow==5.0.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
